requirements.txt
sentence-transformers==2.2.2
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
pymupdf
langchain
langchain-pinecone
langchain_community
langchain_text_splitters
langchain_openai
langchain_experimental

-e .